Settings: simplify all the settings removing a whole old layer (`dev`)
We are not using `settings/dev.py` since we started using
Docker (`settings/docker_compose.py`).
However, we never updated this code.

This commit removes the extra `Dev` class layer by moving the required (and
still used settings) into `settings/test.py` and `settings/docker_compose.py`
since they were inheriting from `settings/dev.py`.
