+++ readthedocs/settings/docker_compose.py
@@ -1,10 +1,10 @@
 import os
 import socket
 
-from .dev import CommunityDevSettings
+from .base import CommunityBaseSettings
 
 
-class DockerBaseSettings(CommunityDevSettings):
+class DockerBaseSettings(CommunityBaseSettings):
 
     """Settings for local development with Docker"""
 
@@ -70,6 +70,9 @@ class DockerBaseSettings(CommunityDevSettings):
 
     RTD_CLEAN_AFTER_BUILD = True
 
+    # Disable password validators on development
+    AUTH_PASSWORD_VALIDATORS = []
+
     @property
     def RTD_EMBED_API_EXTERNAL_DOMAINS(self):
         domains = super().RTD_EMBED_API_EXTERNAL_DOMAINS
@@ -84,6 +87,12 @@ class DockerBaseSettings(CommunityDevSettings):
     @property
     def LOGGING(self):
         logging = super().LOGGING
+
+        logging['handlers']['console']['level'] = 'DEBUG'
+        logging['formatters']['default']['format'] = '[%(asctime)s] ' + self.LOG_FORMAT
+        # Allow Sphinx and other tools to create loggers
+        logging['disable_existing_loggers'] = False
+
         logging['handlers']['console']['formatter'] = 'colored_console'
         logging['loggers'].update({
             # Disable Django access requests logging (e.g. GET /path/to/url)
@@ -185,8 +194,8 @@ class DockerBaseSettings(CommunityDevSettings):
     BUILD_COLD_STORAGE_URL = 'http://storage:9000/builds'
 
     STATICFILES_DIRS = [
-        os.path.join(CommunityDevSettings.SITE_ROOT, 'readthedocs', 'static'),
-        os.path.join(CommunityDevSettings.SITE_ROOT, 'media'),
+        os.path.join(CommunityBaseSettings.SITE_ROOT, 'readthedocs', 'static'),
+        os.path.join(CommunityBaseSettings.SITE_ROOT, 'media'),
     ]
 
     # Remove the checks on the number of fields being submitted
