+++ readthedocs/settings/test.py
@@ -1,10 +1,11 @@
-from __future__ import absolute_import
 import os
 
-from .dev import CommunityDevSettings
+from .base import CommunityBaseSettings
 
 
-class CommunityTestSettings(CommunityDevSettings):
+class CommunityTestSettings(CommunityBaseSettings):
+
+    """Settings for testing environment (e.g. tox)"""
 
     SLUMBER_USERNAME = 'test'
     SLUMBER_PASSWORD = 'test'
@@ -12,6 +13,9 @@ class CommunityTestSettings(CommunityDevSettings):
 
     # A bunch of our tests check this value in a returned URL/Domain
     PRODUCTION_DOMAIN = 'readthedocs.org'
+
+    # Disable password validators on tests
+    AUTH_PASSWORD_VALIDATORS = []
 
     DEBUG = False
     TEMPLATE_DEBUG = False
@@ -26,6 +30,26 @@ class CommunityTestSettings(CommunityDevSettings):
     STRIPE_PUBLISHABLE = 'pk_test_'
     STRIPE_SECRET = 'sk_test_'
 
+    CACHES = {
+        'default': {
+            'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
+            'PREFIX': 'docs',
+        }
+    }
+
+    @property
+    def DATABASES(self):  # noqa
+        return {
+            'default': {
+                'ENGINE': 'django.db.backends.sqlite3',
+                'NAME': os.path.join(self.SITE_ROOT, 'dev.db'),
+            },
+            'telemetry': {
+                'ENGINE': 'django.db.backends.sqlite3',
+                'NAME': os.path.join(self.SITE_ROOT, 'telemetry.dev.db'),
+            },
+        }
+
     @property
     def ES_INDEXES(self):  # noqa - avoid pep8 N802
         es_indexes = super(CommunityTestSettings, self).ES_INDEXES
@@ -36,21 +60,13 @@ class CommunityTestSettings(CommunityDevSettings):
 
     @property
     def LOGGING(self):  # noqa - avoid pep8 N802
-        logging = super(CommunityDevSettings, self).LOGGING
+        logging = super().LOGGING
+
+        logging['handlers']['console']['level'] = 'DEBUG'
+        logging['formatters']['default']['format'] = '[%(asctime)s] ' + self.LOG_FORMAT
+        # Allow Sphinx and other tools to create loggers
+        logging['disable_existing_loggers'] = False
         return logging
 
 
 CommunityTestSettings.load_settings(__name__)
-
-CACHES = {
-    'default': {
-        'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
-        'PREFIX': 'docs',
-    }
-}
-
-if not os.environ.get('DJANGO_SETTINGS_SKIP_LOCAL', False):
-    try:
-        from .local_settings import *  # noqa
-    except ImportError:
-        pass
